literature-manager/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
requests
beautifulsoup4
lxml
PyMySQL
cryptography